// lib/types/site-structure.types.ts
export interface TreeNode extends WebsiteStructure {
  children: TreeNode[];
}